moved components to separate files
